refactor(www): move docs-only helpers into modules/docs
getBaseUrl (was lib/url.ts) and useCopyToClipboard (was
hooks/use-copy-to-clipboard.ts) are consumed only by modules/docs, so
co-locate them with their feature. lib/text.ts stays shared (used by
__root.tsx too).
